# src/main/resources/openapi/loanService.yaml
openapi: 3.0.0
info:
  title: ING Hubs Case Study Loan API
  description: A simple API to manage loans.
  version: 1.0.0

paths:
  /loan:
    post:
      operationId: createLoan
      tags:
        - loan
      parameters:
        - in: query
          name: userId
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateLoanRequest'
      responses:
        200:
          description: Loan successfully created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoanResponse'
  /loan/list:
    get:
      operationId: listLoans
      tags:
        - loan
      parameters:
        - in: query
          name: userId
          schema:
            type: string
        - in: query
          name: numberOfInstallments
          schema:
            type: integer
        - in: query
          name: isPaid
          schema:
            type: boolean
      responses:
        200:
          description: Loans successfully listed
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LoanResponse'
  /loan/installments:
    get:
      operationId: getInstallments
      tags:
        - loan
      parameters:
        - in: query
          name: userId
          schema:
            type: string
        - in: query
          name: loanId
          schema:
            type: string
          required: true
      responses:
        200:
          description: Loans successfully listed
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/InstallmentDetail'
  /loan/pay:
    post:
      operationId: payLoan
      tags:
        - loan
      parameters:
        - in: query
          name: userId
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PayLoanRequest'
      responses:
        200:
          description: Loan is successfully paid

components:
  schemas:
    CreateLoanRequest:
      type: object
      required:
        - amount
        - interestRate
        - numberOfInstallments
      properties:
        amount:
          type: number
        interestRate:
          type: number
          format: double
          minimum: 0.1
          maximum: 0.5
        numberOfInstallments:
          type: integer
    PayLoanRequest:
      type: object
      required:
        - loanId
        - amount
      properties:
        loanId:
          type: string
        amount:
          type: number
    PayLoanResponse:
      type: object
      properties:
        id:
          type: string
        amount:
          type: number
        createDate:
          type: string
        numberOfInstallments:
          type: integer
        isPaid:
          type: boolean
        installments:
          type: array
          items:
            $ref: '#/components/schemas/InstallmentDetail'
    LoanResponse:
      type: object
      properties:
        id:
          type: string
        amount:
          type: number
        createDate:
          type: string
        numberOfInstallments:
          type: integer
        isPaid:
          type: boolean
        installments:
          type: array
          items:
            $ref: '#/components/schemas/InstallmentItem'
    InstallmentItem:
      type: object
      properties:
        dueDate:
          type: string
        amount:
          type: number
        isPaid:
          type: boolean
    InstallmentDetail:
      type: object
      allOf:
        - $ref: '#/components/schemas/InstallmentItem'
      properties:
        paidAmount:
          type: number
        paymentDate:
          type: string